Netjack2 code fixes.

+++ common/JackNetInterface.cpp
@@ -643,7 +643,7 @@ namespace Jack
     bool JackNetSlaveInterface::InitConnection(int time_out_sec)
     {
         jack_log("JackNetSlaveInterface::InitConnection()");
-        uint try_count = (time_out_sec > 0) ? ((1000000 * time_out_sec) / SLAVE_INIT_TIMEOUT) : UINT_MAX;
+        int try_count = (time_out_sec > 0) ? ((1000000 * time_out_sec) / SLAVE_INIT_TIMEOUT) : INT_MAX;
 
         // set the parameters to send
         strcpy(fParams.fPacketType, "params");
@@ -681,7 +681,7 @@ namespace Jack
         return true;
     }
 
-    net_status_t JackNetSlaveInterface::SendAvailableToMaster(long try_count)
+    net_status_t JackNetSlaveInterface::SendAvailableToMaster(int try_count)
     {
         jack_log("JackNetSlaveInterface::SendAvailableToMaster()");
         // utility
